+++ app/api/collectionchildren/[slug]/route.tsx
@@ -2,6 +2,8 @@ import { NextRequest } from "next/server";
 import { fetchApi } from "@/src/utils/fetchApi";
 import { CARDS_PER_PAGE } from "@/src/config/constants";
 
+const oneMonth = 60 * 60 * 24 * 30;
+
 export async function GET(
   req: NextRequest,
   { params }: { params: { slug: string } }
@@ -15,10 +17,10 @@ export async function GET(
     });
   }
   try {
-    const apiUrl = `${process.env.COLLECTIONS_API_URL}/collections/${uuid}/children?page=${page}&perPage=48`;
+    const apiUrl = `${process.env.COLLECTIONS_API_URL}/collections/${uuid}/children?page=${page}&perPage=${CARDS_PER_PAGE}`;
     const response = await fetchApi({
       apiUrl,
-      options: { isRepoApi: false, next: { revalidate: 60 * 60 * 24 * 30 } },
+      options: { isRepoApi: false, next: { revalidate: oneMonth } },
     });
 
     return new Response(JSON.stringify(response), {