+++ app/src/components/pages/searchPage/searchPage.tsx
@@ -20,6 +20,7 @@ import SortMenu from "../../sortMenu/sortMenu";
 import ActiveFilters from "../../search/filters/activeFilters";
 import NoResultsFound from "../../results/noResultsFound";
 import SearchCardGridLoading from "../../grids/searchCardGridLoading";
+import BackToTopLink from "../../backToTopLink/backToTopLink";
 import { SearchResultsType } from "@/src/types/SearchResultsType";
 
 const SearchPage = ({
@@ -190,6 +191,7 @@ const SearchPage = ({
               )} results`}</Heading>
               <SortMenu
                 options={SEARCH_SORT_LABELS}
+                sort={searchResults.sort}
                 searchManager={searchManager}
                 setFiltersExpanded={setFiltersExpanded}
                 updateURL={updateURL}
@@ -239,19 +241,7 @@ const SearchPage = ({
                 flexDir: "column-reverse",
               }}
             >
-              {searchResults.results?.length > 0 && (
-                <Link
-                  minWidth="100px"
-                  isUnderlined={false}
-                  hasVisitedState={false}
-                  gap="xxs"
-                  type="action"
-                  href="#"
-                >
-                  Back to top{"  "}
-                  <Icon name="arrow" iconRotation="rotate180" size="xsmall" />
-                </Link>
-              )}{" "}
+              {searchResults.results?.length > 0 && <BackToTopLink />}{" "}
               <Pagination
                 id="pagination-id"
                 initialPage={searchManager.page}